Modularize logic to UndoRedo component for clarity
